+++ app/toss/fail/page.tsx
@@ -1,7 +1,8 @@
 "use client";
 
-import React from "react";
 import { useSearchParams } from "next/navigation";
+import React from "react";
+
 import "../App.css";
 
 export default function FailPage(): JSX.Element {
